# external/memory/openapi.yaml
...
openapi: 3.1.0
info:
  title: Local Obsidian Memory Service
  summary: Local Obsidian-compatible Markdown memory backend for AI agents.
  description: The configured Markdown Vault is the source of truth. SQLite stores
    disposable search indexes, while optional Git endpoints provide explicit Vault
    versioning.
  version: 0.1.0
paths:
...
  /v1/notes:
    post:
      tags:
      - notes
      summary: Create a note
      description: Creates a Markdown note in the configured Obsidian-compatible Vault.
      operationId: create_note_v1_notes_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteMutationResponse'
        '400':
          description: Invalid request or Git revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Requested note, revision or file was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Operation conflicts with the current Vault or Git state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected database or service error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required service dependency is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/notes/{note_id}/history:
    get:
      tags:
      - notes
      summary: List note history
      operationId: note_history_v1_notes__note_id__history_get
      parameters:
      - name: note_id
        in: path
        required: true
        schema:
          type: string
          title: Note Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteHistoryResponse'
        '400':
          description: Invalid request or Git revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Requested note, revision or file was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Operation conflicts with the current Vault or Git state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected database or service error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required service dependency is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/notes/{note_id}/revisions/{revision}:
    get:
      tags:
      - notes
      summary: Read a historical note revision
      description: Reads the note as it existed at a Git revision, including before
        a rename.
      operationId: revision_v1_notes__note_id__revisions__revision__get
      parameters:
      - name: note_id
        in: path
        required: true
        schema:
          type: string
          title: Note Id
      - name: revision
        in: path
        required: true
        schema:
          type: string
          title: Revision
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevisionResponse'
        '400':
          description: Invalid request or Git revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Requested note, revision or file was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Operation conflicts with the current Vault or Git state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected database or service error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required service dependency is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/notes/{note_id}/restore:
    post:
      tags:
      - notes
      summary: Restore a note revision
      description: Writes a historical revision into the working tree without rewriting
        Git history.
      operationId: restore_v1_notes__note_id__restore_post
      parameters:
      - name: note_id
        in: path
        required: true
        schema:
          type: string
          title: Note Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestoreResponse'
        '400':
          description: Invalid request or Git revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Requested note, revision or file was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Operation conflicts with the current Vault or Git state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected database or service error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required service dependency is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/notes/{note_id}:
    get:
      tags:
      - notes
      summary: Get a note
      operationId: get_note_v1_notes__note_id__get
      parameters:
      - name: note_id
        in: path
        required: true
        schema:
          type: string
          title: Note Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteResponse'
        '400':
          description: Invalid request or Git revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Requested note, revision or file was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Operation conflicts with the current Vault or Git state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected database or service error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required service dependency is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
      - notes
      summary: Update a note
      description: Updates note content and/or metadata; optionally moves its Markdown
        path.
      operationId: update_note_v1_notes__note_id__put
      parameters:
      - name: note_id
        in: path
        required: true
        schema:
          type: string
          title: Note Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteMutationResponse'
        '400':
          description: Invalid request or Git revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Requested note, revision or file was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Operation conflicts with the current Vault or Git state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected database or service error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required service dependency is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - notes
      summary: Delete a note
      operationId: delete_note_v1_notes__note_id__delete
      parameters:
      - name: note_id
        in: path
        required: true
        schema:
          type: string
          title: Note Id
      responses:
        '204':
          description: Note deleted.
        '400':
          description: Invalid request or Git revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Requested note, revision or file was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Operation conflicts with the current Vault or Git state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected database or service error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required service dependency is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
          description: Stable machine-readable error code.
        message:
          type: string
          title: Message
          description: Human-readable error message.
      type: object
      required:
      - code
      - message
      title: ErrorDetail
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      type: object
      required:
      - error
      title: ErrorResponse
...
    GitCommit:
      properties:
        commit:
          type: string
          title: Commit
        timestamp:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Timestamp
        author:
          type: string
          title: Author
        message:
          type: string
          title: Message
      type: object
      required:
      - commit
      - author
      - message
      title: GitCommit
...
    NoteCreate:
      properties:
        path:
          type: string
          title: Path
          description: Vault-relative path ending in `.md`.
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
          description: Optional Obsidian note title.
        content:
          type: string
          title: Content
          description: Markdown body without YAML frontmatter.
          default: ''
        frontmatter:
          additionalProperties: true
          type: object
          title: Frontmatter
          description: Additional YAML frontmatter. `id`, `created` and `updated`
            are managed by the service.
      type: object
      required:
      - path
      title: NoteCreate
    NoteHistoryResponse:
      properties:
        note_id:
          type: string
          title: Note Id
        history:
          items:
            $ref: '#/components/schemas/GitCommit'
          type: array
          title: History
      type: object
      required:
      - note_id
      - history
      title: NoteHistoryResponse
    NoteMutationResponse:
      properties:
        id:
          type: string
          title: Id
        path:
          type: string
          title: Path
      type: object
      required:
      - id
      - path
      title: NoteMutationResponse
    NoteResponse:
      properties:
        id:
          type: string
          title: Id
        path:
          type: string
          title: Path
        title:
          type: string
          title: Title
        content:
          type: string
          title: Content
        raw:
          type: string
          title: Raw
        frontmatter:
          additionalProperties: true
          type: object
          title: Frontmatter
        created:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created
        updated:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated
      type: object
      required:
      - id
      - path
      - title
      - content
      - raw
      - frontmatter
      title: NoteResponse
    NoteUpdate:
      properties:
        path:
          anyOf:
          - type: string
          - type: 'null'
          title: Path
          description: New vault-relative `.md` path.
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
          description: Replacement Obsidian note title.
        content:
          anyOf:
          - type: string
          - type: 'null'
          title: Content
          description: Replacement Markdown body.
        frontmatter:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Frontmatter
          description: Frontmatter fields to merge into the existing note.
      additionalProperties: false
      type: object
      title: NoteUpdate
...
    RestoreRequest:
      properties:
        revision:
          type: string
          maxLength: 200
          minLength: 1
          title: Revision
          description: Git revision to restore.
      type: object
      required:
      - revision
      title: RestoreRequest
    RestoreResponse:
      properties:
        id:
          type: string
          title: Id
        path:
          type: string
          title: Path
        revision:
          type: string
          title: Revision
      type: object
      required:
      - id
      - path
      - revision
      title: RestoreResponse
    RevisionResponse:
      properties:
        note_id:
          type: string
          title: Note Id
        revision:
          type: string
          title: Revision
        path:
          type: string
          title: Path
        raw:
          type: string
          title: Raw
      type: object
      required:
      - note_id
      - revision
      - path
      - raw
      title: RevisionResponse
...